make gmp only on linux
