+++ core/vm/contracts.go
@@ -30,7 +30,6 @@ import (
 	"github.com/consensys/gnark-crypto/ecc/bls12-381/fp"
 	"github.com/consensys/gnark-crypto/ecc/bls12-381/fr"
 	"github.com/holiman/uint256"
-	"github.com/ncw/gmp"
 
 	"github.com/consensys/gnark-crypto/ecc/bn254"
 	"github.com/erigontech/erigon-lib/chain"
@@ -528,61 +527,7 @@ var (
 	errModExpBaseLengthTooLarge     = errors.New("base length is too large")
 	errModExpExponentLengthTooLarge = errors.New("exponent length is too large")
 	errModExpModulusLengthTooLarge  = errors.New("modulus length is too large")
-
-	gmp1 = gmp.NewInt(1)
 )
-
-func (c *bigModExp) Run(input []byte) ([]byte, error) {
-	var (
-		baseLen = new(gmp.Int).SetBytes(getData(input, 0, 32)).Uint64()
-		expLen  = new(gmp.Int).SetBytes(getData(input, 32, 32)).Uint64()
-		modLen  = new(gmp.Int).SetBytes(getData(input, 64, 32)).Uint64()
-	)
-	if c.osaka {
-		// EIP-7823: Set upper bounds for MODEXP
-		if baseLen > 1024 {
-			return nil, errModExpBaseLengthTooLarge
-		}
-		if expLen > 1024 {
-			return nil, errModExpExponentLengthTooLarge
-		}
-		if modLen > 1024 {
-			return nil, errModExpModulusLengthTooLarge
-		}
-	}
-
-	if len(input) > 96 {
-		input = input[96:]
-	} else {
-		input = input[:0]
-	}
-	// Handle a special case when both the base and mod length is zero
-	if baseLen == 0 && modLen == 0 {
-		return []byte{}, nil
-	}
-	// Retrieve the operands and execute the exponentiation
-	var (
-		base = new(gmp.Int).SetBytes(getData(input, 0, baseLen))
-		exp  = new(gmp.Int).SetBytes(getData(input, baseLen, expLen))
-		mod  = new(gmp.Int).SetBytes(getData(input, baseLen+expLen, modLen))
-		v    []byte
-	)
-	switch {
-	case mod.BitLen() == 0:
-		// Modulo 0 is undefined, return zero
-		return common.LeftPadBytes([]byte{}, int(modLen)), nil
-	case base.Cmp(gmp1) == 0:
-		//If base == 1, then we can just return base % mod (if mod >= 1, which it is)
-		v = base.Mod(base, mod).Bytes()
-	//case mod.Bit(0) == 0:
-	//	// Modulo is even
-	//	v = math.FastExp(base, exp, mod).Bytes()
-	default:
-		// Modulo is odd
-		v = base.Exp(base, exp, mod).Bytes()
-	}
-	return common.LeftPadBytes(v, int(modLen)), nil
-}
 
 // newCurvePoint unmarshals a binary blob into a bn256 elliptic curve point,
 // returning it, or an error if the point is invalid.
